+++ kubernetes/modules/redis/versions.tf
@@ -1,0 +1,35 @@
+terraform {
+  required_providers {
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = "~> 2.5.0"
+    }
+
+    helm = {
+      source  = "hashicorp/helm"
+      version = "~> 2.3.0"
+    }
+  }
+  required_version = "~> 1.0.8"
+}
+
+provider "kubernetes" {
+  config_path = "~/.kube/config"
+}
+
+provider "helm" {
+  kubernetes {
+    host                   = var.cluster_endpoint
+    cluster_ca_certificate = base64decode(var.cluster_ca)
+    exec {
+      api_version = "client.authentication.k8s.io/v1alpha1"
+      args = [
+        "eks",
+        "get-token",
+        "--cluster-name",
+        var.cluster_name
+      ]
+      command = "aws"
+    }
+  }
+}