+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add Redis infrastructure and basic BullMQ config
+
 ## v0.1.0
 
 2021-11-18